move data table to account.db
