+++ src/com/seafile/seadroid2/account/AccountDBHelper.java
@@ -2,8 +2,8 @@ package com.seafile.seadroid2.account;
 
 import java.util.List;
 
-import android.database.DatabaseUtils;
 import com.google.common.collect.Lists;
+import com.seafile.seadroid2.data.ServerInfo;
 import com.seafile.seadroid2.provider.AccountNotifier;
 
 import android.content.ContentValues;
@@ -12,8 +12,8 @@ import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
 
-/*
- * Account database helper
+/**
+ * A helper class to manage {@link #DATABASE_NAME} database creation and version management.
  */
 public class AccountDBHelper extends SQLiteOpenHelper {
     // If you change the database schema, you must increment the database version.
@@ -21,11 +21,24 @@ public class AccountDBHelper extends SQLiteOpenHelper {
     private static final String DATABASE_NAME = "account.db";
 
     private static final String ACCOUNT_TABLE_NAME = "Account";
+    private static final String SERVER_INFO_TABLE_NAME = "ServerInfo";
 
     // Account
     private static final String ACCOUNT_COLUMN_SERVER = "server";
     private static final String ACCOUNT_COLUMN_EMAIL = "email";
     private static final String ACCOUNT_COLUMN_TOKEN = "token";
+
+    // Server info
+    private static final String SERVER_INFO_COLUMN_URL = "url";
+    private static final String SERVER_INFO_COLUMN_VERSION = "version";
+    private static final String SERVER_INFO_COLUMN_FEATURE = "feature";
+
+    private static final String SQL_CREATE_SERVER_INFO_TABLE =
+            "CREATE TABLE " + SERVER_INFO_TABLE_NAME + " ("
+                    + SERVER_INFO_COLUMN_URL + " VARCHAR(255) PRIMARY KEY, "
+                    + SERVER_INFO_COLUMN_VERSION + " TEXT NOT NULL, "
+                    + SERVER_INFO_COLUMN_FEATURE + " TEXT NOT NULL" + ")";
+
 
     private static final String SQL_CREATE_ACCOUNT_TABLE =
             "CREATE TABLE " + ACCOUNT_TABLE_NAME + " ("
@@ -53,13 +66,15 @@ public class AccountDBHelper extends SQLiteOpenHelper {
 
     public void onCreate(SQLiteDatabase db) {
         db.execSQL(SQL_CREATE_ACCOUNT_TABLE);
+        db.execSQL(SQL_CREATE_SERVER_INFO_TABLE);
+
     }
 
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
         // This database is only a cache for online data, so its upgrade policy is
         // to simply to discard the data and start over
-        db.execSQL("DELETE TABLE Account;");
-        db.execSQL("DELETE TABLE AccountInfo");
+        db.execSQL("DELETE TABLE IF EXIST " + ACCOUNT_TABLE_NAME + ";");
+        db.execSQL("DELETE TABLE IF EXISTS " + SERVER_INFO_TABLE_NAME + ";");
         onCreate(db);
     }
 
@@ -172,4 +187,43 @@ public class AccountDBHelper extends SQLiteOpenHelper {
         return account;
     }
 
+    public ServerInfo getServerInfo(String url) {
+        String[] projection = {SERVER_INFO_COLUMN_URL, SERVER_INFO_COLUMN_VERSION, SERVER_INFO_COLUMN_FEATURE};
+
+        Cursor c = database.query(SERVER_INFO_TABLE_NAME,
+                projection,
+                "url=?",
+                new String[] {url},
+                null,  // don't group the rows
+                null,  // don't filter by row groups
+                null); // The sort order
+
+        if (!c.moveToFirst()) {
+            c.close();
+            return null;
+        }
+
+        ServerInfo serverInfo = cursorToServerInfo(c);
+
+        c.close();
+        return serverInfo;
+    }
+
+    private ServerInfo cursorToServerInfo(Cursor cursor) {
+        String url = cursor.getString(0);
+        String version = cursor.getString(1);
+        String features = cursor.getString(2);
+        ServerInfo serverInfo = new ServerInfo(url, version, features);
+        serverInfo.setProEdition(features.contains("seafile-pro"));
+        return serverInfo;
+    }
+
+    public void saveServerInfo(ServerInfo serverInfo) {
+        ContentValues values = new ContentValues();
+        values.put(SERVER_INFO_COLUMN_URL, serverInfo.getUrl());
+        values.put(SERVER_INFO_COLUMN_VERSION, serverInfo.getVersion());
+        values.put(SERVER_INFO_COLUMN_FEATURE, serverInfo.getFeatures());
+
+        database.replace(SERVER_INFO_TABLE_NAME, null, values);
+    }
 }
