+++ src/com/seafile/seadroid2/ui/activity/BrowserActivity.java
@@ -40,6 +40,7 @@ import com.actionbarsherlock.view.Window;
 import com.google.common.collect.Lists;
 import com.seafile.seadroid2.*;
 import com.seafile.seadroid2.account.Account;
+import com.seafile.seadroid2.account.AccountDBHelper;
 import com.seafile.seadroid2.account.AccountManager;
 import com.seafile.seadroid2.cameraupload.CameraUploadService;
 import com.seafile.seadroid2.data.*;
@@ -367,7 +368,7 @@ public class BrowserActivity extends SherlockFragmentActivity
     }
 
     private void saveServerInfo(ServerInfo serverInfo) {
-        ServerInfoDBHelper.getInstance().saveServerInfo(serverInfo);
+        AccountDBHelper.getDatabaseHelper(this).saveServerInfo(serverInfo);
     }
 
     /**
@@ -381,7 +382,7 @@ public class BrowserActivity extends SherlockFragmentActivity
         if (account.getServer() == null)
             return false;
 
-        ServerInfo serverInfo = ServerInfoDBHelper.getInstance().getServerInfo(account.getServer());
+        ServerInfo serverInfo = AccountDBHelper.getDatabaseHelper(this).getServerInfo(account.getServer());
         if (serverInfo == null)
             return false;
 
@@ -399,7 +400,7 @@ public class BrowserActivity extends SherlockFragmentActivity
         if (account.getServer() == null)
             return false;
 
-        ServerInfo serverInfo = ServerInfoDBHelper.getInstance().getServerInfo(account.getServer());
+        ServerInfo serverInfo = AccountDBHelper.getDatabaseHelper(this).getServerInfo(account.getServer());
         if (serverInfo == null)
             return false;
